добавлены компоненты презентации мест размещения и презентации отсутствия мест размещения
