+++ project/src/components/places-presentation/no-places-presentation.tsx
@@ -1,0 +1,21 @@
+import CitiesName from '../../types/cities-name';
+
+type NoPlacesPresentationProps = {
+  cityName:CitiesName;
+}
+
+function NoPlacesPresentation({cityName}:NoPlacesPresentationProps):JSX.Element{
+  return(
+    <div className="cities__places-container cities__places-container--empty container">
+      <section className="cities__no-places">
+        <div className="cities__status-wrapper tabs__content">
+          <b className="cities__status">No places to stay available</b>
+          <p className="cities__status-description">We could not find any property available at the moment in {cityName}</p>
+        </div>
+      </section>
+      <div className="cities__right-section"></div>
+    </div>
+  );
+}
+
+export default NoPlacesPresentation;
