+++ project/src/pages/main-screen/main-screen.tsx
@@ -2,15 +2,11 @@ import { useEffect } from 'react';
 import CitiesList from '../../components/cities-list/cities-list';
 import LoadingSpinner from '../../components/loading-spinner/loading-spinner';
 import Logo from '../../components/logo/logo';
-import Map from '../../components/map/map';
 import Navigation from '../../components/navigation/navigation';
-import OffersList from '../../components/offers-list/offers-list';
-import SortForm from '../../components/sort-form/sort-form';
+import PlacesPresentation from '../../components/places-presentation/places-presentation';
 import { useAppDispatch, useAppSelector } from '../../hooks/store-hooks';
 import { isDataLoading } from '../../store/application-data/selectors';
-import { getOffersData } from '../../store/application-process/selectors';
 import { fetchOffers } from '../../store/thunk-actions';
-import { cities, DEFAULT_CITY } from '../../utils/cities';
 
 type MainScreenProps = {
   defaultCardsCount: number;
@@ -20,11 +16,6 @@ type MainScreenProps = {
 function MainScreen({ defaultCardsCount }: MainScreenProps): JSX.Element {
   const dispatch = useAppDispatch();
   const isLoading = useAppSelector(isDataLoading);
-  const {chosenCity, offers} = useAppSelector(getOffersData);
-  const city = cities.find((element) => element.name === chosenCity) ?? DEFAULT_CITY ;
-  const offersCount = offers.length;
-  const offersToShow = offers.slice(0, defaultCardsCount);
-  const points = offersToShow.map((item) => item.location);
 
   useEffect(()=>{
     dispatch(fetchOffers());
@@ -50,17 +41,7 @@ function MainScreen({ defaultCardsCount }: MainScreenProps): JSX.Element {
           <CitiesList/>
         </div>
         <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{offersCount} place{offersCount > 1 ? 's' : ''} to stay in {chosenCity}</b>
-              <SortForm/>
-              <OffersList offers={offersToShow}/>
-            </section>
-            <div className="cities__right-section">
-              <Map className="cities__map" city={city} points={points}/>
-            </div>
-          </div>
+          <PlacesPresentation maxCountToShow={defaultCardsCount}/>
         </div>
       </main>
     </>
